Consolidated logic into 'delegation' vuex module
