+++ app/src/renderer/vuex/getters.js
@@ -5,7 +5,7 @@ export const connected = state => state.node.connected
 export const lastHeader = state => state.node.lastHeader
 
 export const delegates = state => state.delegates
-export const shoppingCart = state => state.shoppingCart.delegates
+export const shoppingCart = state => state.delegation.delegates
 export const user = state => state.user
 export const config = state => state.config
